close the avatar modal after upload refactor names and structure remove route of avatar
