+++ client/src/account/components/Account/index.jsx
@@ -1,19 +1,17 @@
 import React from 'react';
-import axios from 'axios';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 import {withStyles} from '@material-ui/core/styles';
-import Avatar from '@material-ui/core/Avatar';
 import Modal from 'react-responsive-modal';
 import AccountTab from './AccountTab'
 import {Link} from 'react-router-dom';
 import {styles as accountStyles} from '../../AccountStyles';
-import {default as AvatarEditor} from "../Avatar";
+import AvatarUploader from "../AvatarUploader";
 import {connect} from 'react-redux';
 import {compose} from "redux";
 import {getCurrentUserInfo} from '../../actions/authActions';
 import isEmpty from '../../validation/isEmpty'
-import { config } from '../../../config';
+import {config} from '../../../config';
+import {ImageAvatar, LetterAvatar} from "../AvatarUploader/Avatars";
 
 const baseURL = (config.ENV === 'production') ? config.REL_UPLOAD_BASE_URL : config.DEV_UPLOAD_BASE_URL;
 const styles = {
@@ -30,29 +28,6 @@ const styles = {
     },
 };
 
-function ImageAvatars(props) {
-    const {classes, avatarURL, username} = props;
-    return (
-        <div className={classes.row}>
-            <Avatar
-                alt={username}
-                src={avatarURL}
-                className={classNames(classes.avatar, classes.bigAvatar)}
-            />
-        </div>
-    );
-}
-
-function LetterAvatars(props) {
-    const {classes, username} = props;
-    const capital = username ? username.slice(0, 1) : '';
-    return (
-        <div className={classes.row}>
-            <Avatar className={classNames(classes.avatar, classes.bigAvatar)}>{capital}</Avatar>
-        </div>
-    );
-}
-
 function AccountInfo(props) {
     const {username, email} = props;
     return (
@@ -62,7 +37,7 @@ function AccountInfo(props) {
         </div>);
 }
 
-class Account extends React.PureComponent {
+class Account extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
@@ -70,9 +45,6 @@ class Account extends React.PureComponent {
         };
     }
 
-// 		componentWillReceiveProps(nextProps, prevState) {
-// 			this.setState({avatar: nextProps.avatar});
-// }
     componentDidMount() {
         this.props.getCurrentUserInfo();
     };
@@ -97,21 +69,23 @@ class Account extends React.PureComponent {
             avatarType
         } = this.props;
 
-        return <div style={accountStyles.container}>
+        const { container, verticalCenter } = accountStyles;
+
+        return <div style={container}>
             {/* THIS IS UGLY */}
             {isLoggedIn
                 ?
                 (// WHY I HAVE TO SET THE STYLE AGAIN?
-                    <span style={accountStyles.container}>
-				<div style={accountStyles.verticalCenter}>
-					{avatarType === 'letter' ? <LetterAvatars classes={classes} username={username}/> :
-                        <ImageAvatars classes={classes} avatarURL={avatarURL}/>}
+                    <span style={container}>
+				<div style={verticalCenter}>
+					{avatarType === 'letter' ? <LetterAvatar classes={classes} username={username}/> :
+                        <ImageAvatar classes={classes} avatarURL={avatarURL} alt={username}/>}
                     {/* TODO: MAKE IT AS A BANNER ABOVE THE AVATAR */}
                     <button onClick={this.onOpenModal}>Change picture</button>
 				</div>
 				
         <Modal open={avatarPageOpened} onClose={this.onCloseModal} center>
-          <AvatarEditor avatarURL={{avatarURL}}/>
+          <AvatarUploader handleCompletion={this.onCloseModal}/>
         </Modal>
 				
 		<AccountInfo
@@ -126,21 +100,10 @@ class Account extends React.PureComponent {
     }
 }
 
-LetterAvatars.propTypes = {
-    classes: PropTypes.object.isRequired,
-    username: PropTypes.string
-};
-
-ImageAvatars.propTypes = {
-    classes: PropTypes.object.isRequired,
-    avatarURL: PropTypes.string.isRequired
-};
-
 const mapStateToProps = state => {
     // install user info
     if (!isEmpty(state.auth.user)) {
-        const {username, email, avatar: avatarURL} = state.auth.user;
-        // const {imgURL: avatarURL} = state.avatar;
+        const {name: username, email, avatar: avatarURL} = state.auth.user;
         let props = {
             username,
             email,
@@ -149,7 +112,6 @@ const mapStateToProps = state => {
         if (avatarURL) {
             props.avatarType = 'image';
             props.avatarURL = baseURL + avatarURL;
-            console.log('mapStateToProps: avatar', props.avatarURL)
         } else {
             props.avatarType = 'letter';
         }
