+++ client/src/account/components/Register/index.jsx
@@ -7,7 +7,6 @@ import {withRouter} from "react-router-dom";
 import {withStyles} from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 import KFAccountInput from '../KFAccountInput';
-import axios from 'axios';
 
 const styles = theme => ({
   container: {
