+++ client/src/routers/mainRouter.jsx
@@ -10,7 +10,6 @@ import Account from '../account/components/Account';
 import EmailVerification from '../account/components/EmailVerification';
 import VerifyEmail from '../account/utils/verifyEmail';
 import Cart from '../Cart';
-import Avatar from '../account/components/Avatar';
 import Payment from '../Payment/index';
 
 
@@ -27,7 +26,6 @@ const MainRoute = () => (
 			<Route path={'/account'} component={Account} />
 			<Route path={'/verify-email'} component={EmailVerification} />
 			<Route path={'/activate/:token'} component={VerifyEmail}/>
-			<Route path={'/avatar'} component={Avatar}/> {/* modal? */}
 			<Route path={'/payment'} component={Payment}/>
 		</Switch>
 	</Route>
